Ajustando classes

+++ pokedex/src/main/java/com/pokedex/pokedex/service/PokemonService.java
@@ -1,29 +1,21 @@
 package com.pokedex.pokedex.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.pokedex.pokedex.exception.PokemonNotFoundException;
 import com.pokedex.pokedex.mapper.PokemonMapper;
 import com.pokedex.pokedex.model.*;
 import com.pokedex.pokedex.repository.EvolutionRepository;
 import com.pokedex.pokedex.repository.PokemonRepository;
-import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
-import org.springframework.web.client.RestOperations;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
 //Classe que representa as regras de negócio
@@ -38,23 +30,20 @@ public class PokemonService {
     @Autowired
     private EvolutionRepository evolutionRepository;
 
-    @Autowired
-    private RestTemplate restTemplate;
+    public PokemonService(PokeApiService pokeApiService){
+        this.pokeApiService = pokeApiService;
+    }
 
-
-    public EvolutionDetail addNewPokemon(String nameOrNumber) {
+    @Transactional
+    public List<EvolutionDetail> addNewPokemon(String nameOrNumber) {
         // Verificar se o Pokémon já existe na Pokédex
         PokemonResponse pokemonResponses = pokeApiService.getPokemonNameOrNumber(nameOrNumber);
 
-        EvolutionDetail evolutionDetail = PokemonMapper.toDomain(pokemonResponses);
-        for (Pokemon pokemon : evolutionDetail.getEvolution()) {
-            // Verifica se o Pokémon já existe no repositório
-            if (!pokemonRepository.existsById(pokemon.getNumber())) {
-                // Se não existe, salva o Pokémon no repositório
-                pokemonRepository.save(pokemon);
-            }
-        }
-        return evolutionRepository.save(evolutionDetail);
+        List<EvolutionDetail> evolutionDetails = PokemonMapper.toDomain(pokemonResponses);
+
+        evolutionDetails.forEach(evolutionDetail -> evolutionRepository.save(evolutionDetail));
+
+        return evolutionDetails;
     }
 
     public PokemonPageResponse listPokemons(int page, int pageSize) {
@@ -72,10 +61,24 @@ public class PokemonService {
                         pokemonPage.getSize(), (int) pokemonPage.getTotalElements(), pokemonPage.getTotalPages())); // Ajuste conforme necessário para criar a resposta correta
     }
 
-    public boolean deletePokemon(Long number) {
-        Pokemon pokemon = getPokemonByNumber(number);
-        pokemonRepository.delete(pokemon);
-        return true;
+    public void deletePokemonByNameOrNumber(String nameOrNumber) {
+        Pokemon pokemon = findByNameOrNumber(nameOrNumber);
+        if (pokemon!= null){
+            pokemonRepository.delete(pokemon);
+        } else {
+            throw new PokemonNotFoundException("Pokemon not found with name or number: " + nameOrNumber);
+        }
+    }
+
+    private Pokemon findByNameOrNumber(String nameOrNumber){
+        try {
+            long number = Long.parseLong(nameOrNumber);
+            // Se conseguiu converter para número, busca pelo número do Pokémon
+            return pokemonRepository.findByNumber(number);
+        } catch (NumberFormatException e) {
+            // Se não conseguiu converter para número, busca pelo nome do Pokémon
+            return pokemonRepository.findByName(nameOrNumber);
+        }
     }
 
     public Pokemon getPokemonByNumber(Long number) {
